refactor(ai): scope recall under the AI namespace; RecalledLens -> LensCandidate
The recall engine + its projection are AI-search-only today, so they move from the
generic @/lib/lens into the @/lib/ai namespace. With the path carrying the scenario,
the DTO no longer needs to spell it out — RecalledLens becomes LensCandidate (a
candidate the recall surfaced, not yet a recommendation). No behavior change.
